requirements.txt
# Core dependencies
google-adk>=1.0.0

# Data analysis and manipulation
pandas>=1.5.0
numpy>=1.20.0
statistics

# Web scraping and automation
patchright>=1.0.0
playwright>=1.40.0
requests>=2.28.0
curl-cffi>=0.5.0
beautifulsoup4>=4.11.0

# Cryptocurrency trading
ccxt>=4.0.0

# Utilities
urllib3>=1.26.0
backports.zoneinfo>=0.2.1; python_version<"3.9"

# Development and debugging
# logging is built-in, no need to install

# Optional but recommended
python-dotenv>=0.19.0 